# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "AuditEventType": {
        "enum": [
          "tool_call",
          "state_transition",
          "approval",
          "execution",
          "human_interaction"
        ],
        "title": "AuditEventType",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "KnowledgeType": {
        "enum": [
          "business_rule",
          "standard",
          "sop",
          "experience",
          "constraint",
          "team_convention",
          "business_priority"
        ],
        "title": "KnowledgeType",
        "type": "string"
      },
...
      "Page_AuditView_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AuditView"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[AuditView]",
        "type": "object"
      },
      "Page_ChangeEntryView_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ChangeEntryView"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[ChangeEntryView]",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "SessionCookie": {
        "in": "cookie",
        "name": "ops_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Weipai AI Ops Brain",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/audits": {
      "get": {
        "operationId": "audits_api_audits_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Actor"
            }
          },
          {
            "in": "query",
            "name": "event_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AuditEventType"
                },
                {
                  "const": "catalog_edit",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Event Type"
            }
          },
          {
            "in": "query",
            "name": "task_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Task Id"
            }
          },
          {
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_AuditView_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "请先登录"
          },
          "403": {
            "description": "CSRF 或来源校验失败"
          },
          "404": {
            "description": "记录不存在"
          },
          "409": {
            "description": "编辑内容冲突"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "存储或 AI 网关暂时不可用"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Audits",
        "tags": [
          "认知与运营"
        ]
      }
    },
    "/api/audits/{audit_id}": {
      "get": {
        "operationId": "audit_api_audits__audit_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "audit_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Audit Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "请先登录"
          },
          "403": {
            "description": "CSRF 或来源校验失败"
          },
          "404": {
            "description": "记录不存在"
          },
          "409": {
            "description": "编辑内容冲突"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "存储或 AI 网关暂时不可用"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Audit",
        "tags": [
          "认知与运营"
        ]
      }
    },
...
    "/api/changes": {
      "get": {
        "operationId": "changes_api_changes_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "service_name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 256,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Service Name"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "Commit",
                    "Merge",
                    "Build",
                    "Image",
                    "Sync",
                    "Deploy",
                    "Config",
                    "KubernetesEvent",
                    "CloudEvent"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "gitlab",
                    "github",
                    "jenkins",
                    "gitlab_ci",
                    "argocd",
                    "config_center",
                    "kubernetes",
                    "alibaba_cloud"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ChangeEntryView_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "请先登录"
          },
          "403": {
            "description": "CSRF 或来源校验失败"
          },
          "404": {
            "description": "记录不存在"
          },
          "409": {
            "description": "编辑内容冲突"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "存储或 AI 网关暂时不可用"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Changes",
        "tags": [
          "认知与运营"
        ]
      }
    },
    "/api/changes/{change_id}": {
      "get": {
        "operationId": "change_api_changes__change_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "change_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Change Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeEntryView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "请先登录"
          },
          "403": {
            "description": "CSRF 或来源校验失败"
          },
          "404": {
            "description": "记录不存在"
          },
          "409": {
            "description": "编辑内容冲突"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "存储或 AI 网关暂时不可用"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Change",
        "tags": [
          "认知与运营"
        ]
      }
    },
...
    "/api/context-graph/edges": {
      "get": {
        "operationId": "edges_api_context_graph_edges_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "node_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Node Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_EdgeView_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "请先登录"
          },
          "403": {
            "description": "CSRF 或来源校验失败"
          },
          "404": {
            "description": "记录不存在"
          },
          "409": {
            "description": "编辑内容冲突"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "存储或 AI 网关暂时不可用"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Edges",
        "tags": [
          "认知与运营"
        ]
      }
    },
    "/api/context-graph/edges/{edge_id}": {
      "get": {
        "operationId": "edge_api_context_graph_edges__edge_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "edge_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Edge Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EdgeView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "请先登录"
          },
          "403": {
            "description": "CSRF 或来源校验失败"
          },
          "404": {
            "description": "记录不存在"
          },
          "409": {
            "description": "编辑内容冲突"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "存储或 AI 网关暂时不可用"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Edge",
        "tags": [
          "认知与运营"
        ]
      }
    },
    "/api/context-graph/nodes": {
      "get": {
        "operationId": "nodes_api_context_graph_nodes_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_NodeView_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "请先登录"
          },
          "403": {
            "description": "CSRF 或来源校验失败"
          },
          "404": {
            "description": "记录不存在"
          },
          "409": {
            "description": "编辑内容冲突"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "存储或 AI 网关暂时不可用"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Nodes",
        "tags": [
          "认知与运营"
        ]
      }
    },
    "/api/context-graph/nodes/{node_id}": {
      "get": {
        "operationId": "node_api_context_graph_nodes__node_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Node Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "请先登录"
          },
          "403": {
            "description": "CSRF 或来源校验失败"
          },
          "404": {
            "description": "记录不存在"
          },
          "409": {
            "description": "编辑内容冲突"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "存储或 AI 网关暂时不可用"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Node",
        "tags": [
          "认知与运营"
        ]
      }
    },
...
    "/api/knowledge": {
      "get": {
        "operationId": "knowledge_api_knowledge_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/KnowledgeType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_KnowledgeView_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "请先登录"
          },
          "403": {
            "description": "CSRF 或来源校验失败"
          },
          "404": {
            "description": "记录不存在"
          },
          "409": {
            "description": "编辑内容冲突"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "存储或 AI 网关暂时不可用"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Knowledge",
        "tags": [
          "认知与运营"
        ]
      },
      "post": {
        "operationId": "create_knowledge_api_knowledge_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "title": "X-Csrf-Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "请先登录"
          },
          "403": {
            "description": "CSRF 或来源校验失败"
          },
          "404": {
            "description": "记录不存在"
          },
          "409": {
            "description": "编辑内容冲突"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "存储或 AI 网关暂时不可用"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Create Knowledge",
        "tags": [
          "认知与运营"
        ]
      }
    },
...
  }
}